requirements.txt
numpy-stl>=2.0.0
numpy>=1.20.0
PyQt5>=5.15.0
